find a bug, could not receive from uart

+++ Components/Chassis/chassis.h
@@ -1,6 +1,8 @@
 #ifndef __CHASSIS_H
 #define __CHASSIS_H
 
+#include <stdint.h>
+
 /// CHASSIS_RW = CHASSIS_RA + CHASSIS_RB
 #define CHASSIS_RW 141.f   // 单位 mm
 #define WHEEL_RADIUS 38.f  // 单位 mm
@@ -10,11 +12,11 @@ void chassis_init(void);
 
 /// 控制相对位置, 单位为 mm
 /// !!! 不要在中断里使用
-void chassis_control_dest(float x, float y);
+void chassis_control_dest(uint16_t x, uint16_t y);
 
 /// 控制相对角度
 /// !!! 不要在中断里使用
-void chassis_control_angu(float w);
+void chassis_control_angu(int8_t w);
 
 /// 电机事件回调函数, 控制速度更新频率
 /// ```
